fix: ajustes nos DDL's para criarem os schemas flyway de maneira correta desde a primeira execução

+++ src/main/resources/db/migration/V1__create_table_pedido.sql
@@ -1,5 +1,5 @@
 CREATE TABLE pedido (
-    id_pedido INT AUTO_INCREMENT PRIMARY KEY,
+    id_pedido BIGINT AUTO_INCREMENT PRIMARY KEY,
     id_cliente INT NOT NULL,
     data_pedido DATETIME NOT NULL,
     id_status_pedido INT NOT NULL,
